Feature: Readme.md
Added the Readme.md file with some information (at the moment) of all
the functions that this app will have or has already at this point of
the development. At the end of the project, when the final version comes
out, there will be an instruction of all the steps for launching it, and
how to preview the content of the already said application.